homework5

+++ src/main/java/ua/hillel/nemirovets/homework/homework5/Homework5.java
@@ -11,8 +11,8 @@ public class Homework5 {
         int minWarrior = 9;
         int minBowman = 15;
         int minRider = 12;
-        double countMinWarrior = countLiWarrior* 1.5;
-        int minGeneralAttack = (minWarrior + minBowman + minRider) * (int) countMinWarrior;
+        int countMinWarrior =  (int) (countLiWarrior* 1.5);
+        int minGeneralAttack = (minWarrior + minBowman + minRider) *  countMinWarrior;
 
         System.out.println("General Attack Li = " + liGeneralAttack);
         System.out.println("General Attack Min = " +minGeneralAttack);
